Changed the video setup to another file
