+++ SocialNetworkingApp/Explore/ExplorePostViewController.swift
@@ -31,7 +31,7 @@ class ExplorePostViewController: UIViewController {
         let cells = explorePostTableView.visibleCells
         for cell in cells {
             if let cell = cell as? PostTableViewCell {
-                cell.player?.play()
+                cell.postVideoView.player?.play()
             }
         }
     }
@@ -41,7 +41,7 @@ class ExplorePostViewController: UIViewController {
         let cells = explorePostTableView.visibleCells
         for cell in cells {
             if let cell = cell as? PostTableViewCell {
-                cell.player?.pause()
+                cell.postVideoView.player?.pause()
             }
         }
     }
@@ -88,13 +88,13 @@ extension ExplorePostViewController: UITableViewDataSource{
 extension ExplorePostViewController: UITableViewDelegate{
     func tableView(_ tableView: UITableView, didEndDisplaying cell: UITableViewCell, forRowAt indexPath: IndexPath) {
         if let cell = cell as? PostTableViewCell {
-            cell.player?.pause()
+            cell.postVideoView.player?.pause()
         }
     }
     
     func tableView(_ tableView: UITableView, willDisplay cell: UITableViewCell, forRowAt indexPath: IndexPath) {
         if let cell = cell as? PostTableViewCell{
-            cell.player?.play()
+            cell.postVideoView.player?.play()
         }
     }
 }
